+++ src/cleaner.rs
@@ -73,14 +73,17 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
     // Do not remove any attributes before this methods
     // since they uses them.
 
+    // independent task, doesn't impact any other tasks
     if options.remove_title {
         remove_element(doc, ElementId::Title);
     }
 
+    // independent task, doesn't impact any other tasks
     if options.remove_desc {
         remove_element(doc, ElementId::Desc);
     }
 
+    // independent task, doesn't impact any other tasks
     if options.remove_metadata {
         remove_element(doc, ElementId::Metadata);
     }
@@ -89,13 +92,14 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
         remove_unused_defs(doc);
     }
 
+    // impact only `linearGradient` and `radialGradient`
     if options.remove_invalid_stops {
         remove_invalid_stops(doc);
     }
 
     if options.apply_transform_to_gradients {
-        // Apply transform to gradients before processing to simplify duplicates
-        // detecting and merging.
+        // apply transform to gradients before processing to simplify duplicates
+        // detecting and merging
         apply_transforms::apply_transform_to_gradients(doc);
     }
 
@@ -116,15 +120,22 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
     }
 
     if options.apply_transform_to_gradients {
-        // Do it again, because something may changed after gradient processing.
+        // do it again, because something may changed after gradients processing
         apply_transforms::apply_transform_to_gradients(doc);
     }
 
+    // run before `apply_transform_to_shapes` and `process_paths`,
+    // because it impact transform processing
+    if options.ungroup_groups {
+        ungroup_groups(doc, options);
+    }
+
+    // run before 'convert_shapes_to_paths'
     if options.apply_transform_to_shapes {
-        // Apply before 'convert_shapes_to_paths'.
         apply_transforms::apply_transform_to_shapes(doc);
     }
 
+    // impact only shapes
     if options.convert_shapes {
         convert_shapes_to_paths(doc);
     }
@@ -140,12 +151,14 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
         remove_invisible_elements(doc);
     }
 
+    // impact only `linearGradient` and `radialGradient`
     if options.regroup_gradient_stops {
         regroup_gradient_stops(doc);
     }
 
+    // ungroup again
     if options.ungroup_groups {
-        ungroup_groups(doc);
+        ungroup_groups(doc, options);
     }
 
     if options.resolve_use {
@@ -174,8 +187,13 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
         remove_unused_coordinates(doc);
     }
 
-    // Run only after attributes processed, because
-    // there is no point in grouping default/unneeded attributes.
+    // ungroup again
+    if options.ungroup_groups {
+        ungroup_groups(doc, options);
+    }
+
+    // run only after attributes processed, because
+    // there is no point in grouping default/unneeded attributes
     if options.group_by_style {
         group_by_style(doc);
     }
@@ -183,18 +201,22 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
     // final fixes
     // list of things that can't break anything
 
+    // independent task, doesn't impact any other tasks
     if options.remove_unreferenced_ids {
         remove_unreferenced_ids(doc);
     }
 
+    // independent task, doesn't impact any other tasks
     if options.trim_ids {
         trim_ids(doc);
     }
 
+    // independent task, doesn't impact any other tasks
     if options.remove_version {
         remove_version(doc);
     }
 
+    // run at last, because it can remove `defs` element which is used by many algorithms
     if options.ungroup_defs {
         ungroup_defs(doc);
     }
@@ -202,7 +224,7 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
     remove_empty_defs(doc);
     fix_xmlns_attribute(doc, options.remove_xmlns_xlink_attribute);
 
-    // NOTE: must be run at last, since it breaks the linking.
+    // NOTE: must be run at last, since it breaks the linking
     join_style_attributes(doc, options.join_style_attributes, opt);
 
     Ok(())