+++ src/task/apply_transforms/preproc.rs
@@ -1,0 +1,175 @@
+/****************************************************************************
+**
+** svgcleaner could help you to clean up your SVG files
+** from unnecessary data.
+** Copyright (C) 2012-2017 Evgeniy Reizner
+**
+** This program is free software; you can redistribute it and/or modify
+** it under the terms of the GNU General Public License as published by
+** the Free Software Foundation; either version 2 of the License, or
+** (at your option) any later version.
+**
+** This program is distributed in the hope that it will be useful,
+** but WITHOUT ANY WARRANTY; without even the implied warranty of
+** MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+** GNU General Public License for more details.
+**
+** You should have received a copy of the GNU General Public License along
+** with this program; if not, write to the Free Software Foundation, Inc.,
+** 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+**
+****************************************************************************/
+
+use task::short::{EId, AId};
+use super::utils;
+use options::Options;
+
+use svgdom::{Document, Node};
+use svgdom::types::Transform;
+
+// If group has transform and contains only valid elements
+// we can apply the group's transform to children before applying transform to
+// actual elements.
+pub fn prepare_transforms(doc: &Document, options: &Options) {
+    let mut valid_elems: Vec<EId> = Vec::with_capacity(6);
+    if options.apply_transform_to_shapes {
+        valid_elems.extend_from_slice(&[EId::Rect, EId::Circle, EId::Ellipse, EId::Line]);
+    }
+
+    if options.paths_to_relative && options.apply_transform_to_paths {
+        valid_elems.push(EId::Path);
+    }
+
+    if valid_elems.is_empty() {
+        // nothing to do
+        return;
+    }
+
+    let iter = doc.descendants().svg().filter(|n|    n.is_tag_name(EId::G)
+                                                  && n.has_attribute(AId::Transform));
+
+    for node in iter {
+        if !utils::has_valid_transform(&node) || !utils::is_valid_attrs(&node) {
+            continue;
+        }
+
+        // check that all children is valid
+        let is_valid = node.children().svg().all(|n| {
+            let tag_name = n.tag_id().unwrap();
+
+            let is_valid_coords = if tag_name == EId::Path {
+                true
+            } else {
+                utils::is_valid_coords(&n)
+            };
+
+               valid_elems.contains(&tag_name)
+            && utils::has_valid_transform(&n)
+            && utils::is_valid_attrs(&n)
+            && is_valid_coords
+        });
+
+        if is_valid {
+            let ts = utils::get_ts(&node);
+            apply_ts_to_children(&node, ts);
+            node.remove_attribute(AId::Transform);
+        }
+    }
+}
+
+fn apply_ts_to_children(node: &Node, ts: Transform) {
+    for child in node.children().svg() {
+        if child.has_attribute(AId::Transform) {
+            // we should multiply transform matrices
+            let mut ts1 = ts;
+            let ts2 = utils::get_ts(&child);
+            ts1.append(&ts2);
+            child.set_attribute(AId::Transform, ts1);
+        } else {
+            child.set_attribute(AId::Transform, ts);
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use svgdom::{Document, WriteToString};
+
+    macro_rules! test {
+        ($name:ident, $in_text:expr, $out_text:expr) => (
+            #[test]
+            fn $name() {
+                let doc = Document::from_str($in_text).unwrap();
+
+                let mut options = Options::default();
+                options.apply_transform_to_shapes = true;
+                options.paths_to_relative = true;
+                options.apply_transform_to_paths = true;
+
+                prepare_transforms(&doc, &options);
+
+                let mut opt = write_opt_for_tests!();
+                opt.simplify_transform_matrices = true;
+                assert_eq_text!(doc.to_string_with_opt(&opt), $out_text);
+            }
+        )
+    }
+
+    test!(apply_1,
+"<svg>
+    <g transform='translate(10 20)'>
+        <rect transform='scale(2)'/>
+        <path/>
+    </g>
+</svg>",
+"<svg>
+    <g>
+        <rect transform='matrix(2 0 0 2 10 20)'/>
+        <path transform='translate(10 20)'/>
+    </g>
+</svg>
+");
+
+    test_eq!(keep_1,
+"<svg>
+    <g transform='scale(10 20)'>
+        <path d='M 10 20 L 30 40'/>
+    </g>
+    <g transform='scale(10 20)'>
+        <rect d='M 10 20 L 30 40'/>
+    </g>
+</svg>
+");
+
+    test_eq!(keep_2,
+"<svg>
+    <linearGradient id='lg1'/>
+    <g fill='url(#lg1)'>
+        <rect/>
+    </g>
+    <g fill='url(#lg1)'>
+        <path/>
+    </g>
+</svg>
+");
+
+    // group should containt only supported children
+    test_eq!(keep_3,
+"<svg>
+    <g transform='scale(10 20)'>
+        <g/>
+        <rect/>
+    </g>
+</svg>
+");
+
+    // both transforms should be valid
+    test_eq!(keep_4,
+"<svg>
+    <g transform='scale(10)'>
+        <rect transform='scale(10 30)'/>
+    </g>
+</svg>
+");
+}