+++ src/task/apply_transforms/shapes.rs
@@ -27,58 +27,6 @@ use svgdom::{Document, Node, Attributes};
 use svgdom::types::Transform;
 
 pub fn apply_transform_to_shapes(doc: &Document) {
-    // If group has transform and contains only valid shapes
-    // we can apply the group's transform to children before applying transform to
-    // actual shape's coordinates.
-    //
-    // We use own implementation, because 'task::ungroup_groups' method does not support it.
-    let iter = doc.descendants().svg().filter(|n|    n.is_tag_name(EId::G)
-                                                  && n.has_attribute(AId::Transform));
-
-    for node in iter {
-        if !utils::has_valid_transform(&node) || !utils::is_valid_attrs(&node) {
-            continue;
-        }
-
-        // check that all children is valid
-        let is_valid = node.children().svg().all(|n| {
-            let flag = match n.tag_id().unwrap() {
-                  EId::Rect
-                | EId::Circle
-                | EId::Ellipse
-                | EId::Line => true,
-                _ => false,
-            };
-
-               flag
-            && utils::has_valid_transform(&n)
-            && utils::is_valid_attrs(&n)
-            && utils::is_valid_coords(&n)
-        });
-
-        if is_valid {
-            let ts = utils::get_ts(&node);
-
-            // apply group's transform to children
-            for child in node.children().svg() {
-                if child.has_attribute(AId::Transform) {
-                    // we should multiply transform matrices
-                    let mut ts1 = ts;
-                    let ts2 = utils::get_ts(&child);
-                    ts1.append(&ts2);
-                    child.set_attribute(AId::Transform, ts1);
-                } else {
-                    child.set_attribute(AId::Transform, ts);
-                }
-            }
-
-            node.remove_attribute(AId::Transform);
-
-            // we do not remove group element here, because
-            // it's 'task::ungroup_groups' job
-        }
-    }
-
     // apply transform to shapes
     let iter = doc.descendants().svg().filter(|n| n.has_attribute(AId::Transform));
     for node in iter {
@@ -112,7 +60,7 @@ fn process<F>(node: &Node, func: F)
     if ts.has_scale() {
         // we must update 'stroke-width' if transform had scale part in it
         let (sx, _) = ts.get_scale();
-        ::task::utils::recalc_stroke_width(node, sx);
+        ::task::utils::recalc_stroke(node, sx);
     }
 }
 
@@ -240,23 +188,6 @@ mod tests {
 </svg>",
 "<svg>
     <line stroke-width='2' x1='30' x2='30' y1='50' y2='50'/>
-</svg>
-");
-
-    test!(apply_g_1,
-"<svg>
-    <g transform='translate(10 20) scale(2)'>
-        <rect height='10' width='10' x='10' y='10' transform='scale(2)'/>
-        <rect height='10' width='10' x='10' y='10'/>
-        <rect height='10' width='10' x='10' y='10'/>
-    </g>
-</svg>",
-"<svg>
-    <g>
-        <rect height='40' stroke-width='4' width='40' x='50' y='60'/>
-        <rect height='20' stroke-width='2' width='20' x='30' y='40'/>
-        <rect height='20' stroke-width='2' width='20' x='30' y='40'/>
-    </g>
 </svg>
 ");